cafe_manager/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
pytest==8.3.4
requests==2.32.3
serializers==0.2.4
sqlparse==0.5.3
urllib3==2.3.0

